Separate out domains module in subspace-runtime
